refactor to avoid using ids in the search inputs
